+++ CHANGELOG.md
@@ -42,6 +42,8 @@
 
 ## [1.1.20](https://www.npmjs.com/package/vitessce/v/1.1.20) - 2022-04-21
 
+
+
 ### Added
 - Add a tutorial that describes how to deploy a Vitessce web app to GitHub pages.
 - Support `var` alias for AnnData to display altenrative gene names via new `geneAlias` field.